# [SPARK-37326][SQL] Support TimestampNTZ in CSV data source