--- a/sql/catalyst/src/main/scala/org/apache/spark/sql/catalyst/csv/CSVInferSchema.scala
+++ b/sql/catalyst/src/main/scala/org/apache/spark/sql/catalyst/csv/CSVInferSchema.scala
@@ -178,8 +178,7 @@ class CSVInferSchema(val options: CSVOptions) extends Serializable {
     // We can only parse the value as TimestampNTZType if it does not have zone-offset or
     // time-zone component and can be parsed with the timestamp formatter.
     // Otherwise, it is likely to be a timestamp with timezone.
-    if ((allCatch opt !timestampNTZFormatter.isTimeZoneSet(field)).getOrElse(false) &&
-        (allCatch opt timestampNTZFormatter.parseWithoutTimeZone(field)).isDefined) {
+    if ((allCatch opt timestampNTZFormatter.parseWithoutTimeZone(field)).isDefined) {
       SQLConf.get.timestampType
     } else {
       tryParseTimestamp(field)