--- a/sql/catalyst/src/main/scala/org/apache/spark/sql/catalyst/csv/UnivocityParser.scala
+++ b/sql/catalyst/src/main/scala/org/apache/spark/sql/catalyst/csv/UnivocityParser.scala
@@ -204,9 +204,6 @@ class UnivocityParser(
 
     case _: TimestampNTZType => (d: String) =>
       nullSafeDatum(d, name, nullable, options) { datum =>
-        if (timestampNTZFormatter.isTimeZoneSet(datum)) {
-          throw QueryExecutionErrors.cannotParseStringAsDataTypeError(name, datum, TimestampNTZType)
-        }
         timestampNTZFormatter.parseWithoutTimeZone(datum)
       }