--- a/sql/catalyst/src/main/scala/org/apache/spark/sql/catalyst/util/TimestampFormatter.scala
+++ b/sql/catalyst/src/main/scala/org/apache/spark/sql/catalyst/util/TimestampFormatter.scala
@@ -165,13 +165,7 @@ class Iso8601TimestampFormatter(
   }
 
   override def format(localDateTime: LocalDateTime): String = {
-    // If the legacy time parser policy is selected, we can only write timestamp with timezone,
-    // we will use the default time zone for it.
-    if (SQLConf.get.legacyTimeParserPolicy == LEGACY) {
-      format(toJavaTimestamp(instantToMicros(localDateTime.atZone(zoneId).toInstant)))
-    } else {
-      localDateTime.format(formatter)
-    }
+    localDateTime.format(formatter)
   }
 
   override def validatePatternString(checkLegacy: Boolean): Unit = {