--- a/sql/catalyst/src/main/scala/org/apache/spark/sql/catalyst/util/TimestampFormatter.scala
+++ b/sql/catalyst/src/main/scala/org/apache/spark/sql/catalyst/util/TimestampFormatter.scala
@@ -31,9 +31,10 @@ import org.apache.spark.sql.catalyst.util.DateTimeConstants._
 import org.apache.spark.sql.catalyst.util.DateTimeUtils._
 import org.apache.spark.sql.catalyst.util.LegacyDateFormats.{LegacyDateFormat, LENIENT_SIMPLE_DATE_FORMAT}
 import org.apache.spark.sql.catalyst.util.RebaseDateTime._
+import org.apache.spark.sql.errors.QueryExecutionErrors
 import org.apache.spark.sql.internal.SQLConf
 import org.apache.spark.sql.internal.SQLConf.LegacyBehaviorPolicy._
-import org.apache.spark.sql.types.Decimal
+import org.apache.spark.sql.types.{Decimal, TimestampNTZType}
 import org.apache.spark.unsafe.types.UTF8String
 
 sealed trait TimestampFormatter extends Serializable {
@@ -69,19 +70,6 @@ sealed trait TimestampFormatter extends Serializable {
   def parseWithoutTimeZone(s: String): Long =
     throw new IllegalStateException(
       s"The method `parseWithoutTimeZone(s: String)` should be implemented in the formatter " +
-        "of timestamp without time zone")
-
-  /**
-   * Returns true if the parsed timestamp contains the time zone component, false otherwise.
-   * Used to determine if the timestamp can be inferred as timestamp without time zone.
-   *
-   * @param s - string with timestamp to inspect
-   * @return whether the timestamp string has the time zone component defined.
-   */
-  @throws(classOf[IllegalStateException])
-  def isTimeZoneSet(s: String): Boolean =
-    throw new IllegalStateException(
-      s"The method `isTimeZoneSet(s: String)` should be implemented in the formatter " +
         "of timestamp without time zone")
 
   def format(us: Long): String
@@ -134,16 +122,14 @@ class Iso8601TimestampFormatter(
   override def parseWithoutTimeZone(s: String): Long = {
     try {
       val parsed = formatter.parse(s)
+      val parsedZoneId = parsed.query(TemporalQueries.zone())
+      if (parsedZoneId != null) {
+        throw QueryExecutionErrors.cannotParseStringAsDataTypeError(pattern, s, TimestampNTZType)
+      }
       val localDate = toLocalDate(parsed)
       val localTime = toLocalTime(parsed)
       DateTimeUtils.localDateTimeToMicros(LocalDateTime.of(localDate, localTime))
     } catch checkParsedDiff(s, legacyFormatter.parse)
-  }
-
-  override def isTimeZoneSet(s: String): Boolean = {
-    val parsed = formatter.parse(s)
-    val parsedZoneId = parsed.query(TemporalQueries.zone())
-    parsedZoneId != null
   }
 
   override def format(instant: Instant): String = {
@@ -207,13 +193,14 @@ class DefaultTimestampFormatter(
 
   override def parseWithoutTimeZone(s: String): Long = {
     try {
-      DateTimeUtils.stringToTimestampWithoutTimeZoneAnsi(UTF8String.fromString(s))
+      val utf8Value = UTF8String.fromString(s)
+      val (_, zoneIdOpt, _) = DateTimeUtils.parseTimestampString(utf8Value)
+      if (zoneIdOpt.isDefined) {
+        throw QueryExecutionErrors.cannotParseStringAsDataTypeError(
+          TimestampFormatter.defaultPattern(), s, TimestampNTZType)
+      }
+      DateTimeUtils.stringToTimestampWithoutTimeZoneAnsi(utf8Value)
     } catch checkParsedDiff(s, legacyFormatter.parse)
-  }
-
-  override def isTimeZoneSet(s: String): Boolean = {
-    val (_, zoneIdOpt, _) = parseTimestampString(UTF8String.fromString(s))
-    zoneIdOpt.isDefined
   }
 }