@sadikovi 2021-11-19
IMHO, we should revert to the original proposal with `isTimeZoneSet` method: it preserves the original API at the expense of having an additional method to check timezone.