@sadikovi 2021-11-22
@cloud-fan @gengliangwang I updated the code, followed the advice to pass an extra boolean flag `ignoreTimeZone` in the relevant methods. Can you review the changes? Thanks!